Pause after X seconds feature
Sometime, we need to do action on the pen at regular interval (like shaking a posca pen).
This patch define a UI field to ask for a pause after X seconds.
Then the plot will stop after X seconds.
After a resume, the plot will honor any change to the UI field, and stop again after X seconds, and so on.

This should allow to do any regular operation on a pen, without the need to stay near the plot to press pause at the right time.